mir: remove objects
These were a middle layer that really had no reason to exist, so just be
rid of them.
