+++ src/mir/mir.cpp
@@ -60,4 +60,48 @@ bool Variable::operator==(const Variable & other) const {
     return name == other.name && version == other.version;
 }
 
+const bool File::is_built() const { return built; }
+
+std::string File::get_name() const { return name; }
+
+std::filesystem::path File::relative_to_source_dir() const {
+    if (built) {
+        std::error_code ec{};
+        auto p = std::filesystem::relative(build_root / subdir / name, source_root / subdir, ec);
+        if (ec) {
+            // TODO: better error handling
+            throw Util::Exceptions::MesonException{"Failed to create relative path"};
+        }
+        return p;
+    } else {
+        return subdir / name;
+    }
+}
+
+std::filesystem::path File::relative_to_build_dir() const {
+    if (!built) {
+        std::error_code ec{};
+        auto p = std::filesystem::relative(source_root / subdir / name, build_root / subdir, ec);
+        if (ec) {
+            // TODO: better error handling
+            throw Util::Exceptions::MesonException{"Failed to create relative path"};
+        }
+        return p;
+    } else {
+        return subdir / name;
+    }
+}
+
+bool File::operator==(const File & f) const {
+    return subdir / name == f.subdir / f.name && built == f.built;
+}
+
+bool File::operator!=(const File & f) const {
+    return subdir / name != f.subdir / f.name || built != f.built;
+}
+
+std::ostream & operator<<(std::ostream & os, const File & f) {
+    return os << (f.is_built() ? f.build_root : f.source_root) / f.subdir / f.get_name();
+}
+
 } // namespace MIR
